Commit Log: effectively disable strict threshold maintenance behavior
This commit effectively disables the commitlog behavior that is related
to strict threshold maintenance.
The first change is to stop accounting the wasted bytes (or file slack bytes)
as there is at least one bug in that area that puts the whole system out
of balance. The second, makes the segment flash behave as it behaved
before and try to flush all segments up to the active one instead of
evicting exactly the segments that will make the commit log go below the
disk usage threshold.
The changes makes the commitlog behave closer to the way it used to
behave before strict disk usage was introduced. This is meant to be
temporary and allow for gradual re-enabling of the code once the bugs or
the source of unpredictable behavior is found.

Refs scylladb/scylla#9331