schema_registry: schema_registry_entry: prepare for multiple attached schemas
The same schema registry entry can be associated with multiple schema
instances. Remember that a schema registry entry represents a given
schema version, and a given schema version can have multiple schema
objects associated with it. No more proof is needed for this then the
existence of a copy constructor for schema.