schema_builder: add constructors variants with a registry parameter
So callers can propagate a registry instance to the to-be-created
schema. The constructor taking a schema instance obtains the registry
from this schema if it has a registry entry.
This registry instance is not used yet. It will be passed to the schema
in the next patches, after some further preparatory steps.