schema_registry: be more lenient with tests
The schema registry uses a timer to evict unused schema registry
entires when the last instance of the associated schema is destroyed.
This timer requires a reactor instance to be present to work. The most
simple tests that only use the boost framework but don't setup a seastar
reactor don't have this. To support the existence of such tests, we
check for the reactor's existence before creating the timer.
This check will not disturb production code: we will always have a
reactor instance in production.