schema: self-register when constructed
If a non-null registry instance is passed as the new registry argument
of the constructor.
After a gradual transition of the code to a point where a registry
instance is propagated to all schema construction sites, we will make
this argument mandatory at which point all schemas will self-register on
construction.