test/alternator: improve basic test for TTL expiration
This patch improves test_ttl.py::test_ttl_expiration in two ways:

First, it checks yet another case - that items that have the wrong type
for the expiration-time column (e.g., a string) never get expired - even
if that string happens to contain a number that looks like an expiration
time.

Second, instead of the huge 15-minute duration for this test, the
test now has a configurable duration; We still need to use a very long
duration on AWS, but in Scylla we expect to be able to configure the
TTL scan frequency, and can finish this test in just a few seconds!
We already have experimental code which makes this test pass in just
3 seconds.

Signed-off-by: Nadav Har'El <nyh@scylladb.com>