cql3: deprecate and start replacing cql3::util::maybe_quote
cql3::util::maybe_quote() is a utility function formatting an identifier
name (table name, column name, etc.) that needs to be embedded in a CQL
statement - and might require quoting if it contains non-alphanumeric
characters, uppercase characters, or a CQL keyword.

maybe_quote() made a valiant effort to only quote the identifier name
if neccessary - e.g., a lowercase name does not need quoting. Or, so
we thought. I now consider that effort to be misdirected: It turns out
that even lowercase names sometimes need to be quoted - when they are
CQL keywords they cannot be used without quoting. This can cause
problems for code that wants to generate CQL statements, such as the
materialized-view problem in issue #9450.

So this patch concedes that maybe_quote() is problematic - it doesn't
always quote when necessary - and deprecates it. Instead it adds a new
function - quote() - which unconditionally quotes an identifier name.
It is always legal to add quotes even when not necessary.

I did not want to replace all uses of maybe_quote() with quote() because
I am not sure how to test all the places it is used. So I changed only
the one place in the code which is needed to fix issue #9450 - the
column identifier's to_cql_string() function. Probably eventually all
other calls to maybe_quote() should be reviewed, tests added for the
CQL-keyword case, and the call to maybe_quote() should be replaced by
quote().

Fixes #9450

Signed-off-by: Nadav Har'El <nyh@scylladb.com>