gossip: Make gossip listener failure handling more robust
Since commit 0ed84710d9e, a failure in the gossip listener callbacks or
replication of gossip app state would cause a whole scylla process
to abort.

This is very dangerous if multiple nodes in the cluster have similar
situation and abort at the same time. We recently saw a case where the
host_id of a bootstrapping node was not available in gossip because the
bootstrapping node was removed from gossip since it was not updating
gossip status for some time. But later, the bootstrapping node finished
streaming and advertised NORMAL status. The gossip listener throwed
because the host_id could not be found. As a result, multiple nodes
aborted at the same time.

To improve this, in case of gossip listener failure, instead of abort
immediately, we isolate the node in question. User has to run a rolling
restart to recover.

Fixes #9657