repair: Use local gossiper reference
There are two places in repair that call for global gossiper instance.
However, the repair_service already has sharded gossiper on board, and
it can use it directly in the first place.

The second place is called from inside repair_info method. This place
is fixed by keeping the gossiper reference on the info, just like it's
done for other services that info needs.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>